requirements.txt
boto==2.49.0
catboost==0.26
cloudpickle==1.6.0
cycler==0.10.0
graphviz==0.17
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.4.2
numba==0.53.1
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
plotly==5.1.0
psycopg2==2.9.1
pylab-sdk==1.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
shap==0.39.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
tenacity==8.0.1
threadpoolctl==2.2.0
tqdm==4.62.0